🎉 Add preview functionality for selected steps in dashboard
