+++ apps/wizard/app_pages/dashboard/utils.py
@@ -1,3 +1,5 @@
+from typing import List
+
 import pandas as pd
 import streamlit as st
 
@@ -194,3 +196,18 @@ def load_steps_df_to_display(show_all_channels: bool, reload_key: int) -> pd.Dat
     df = df.astype(dtypes_new)
     df["date_of_next_update"] = pd.to_datetime(df["date_of_next_update"], errors="coerce")
     return df
+
+
+def _add_steps_to_operations(steps_related: List[str]):
+    """Add steps to the operations list."""
+    # Remove those already in operations list
+    new_selected_steps = [step for step in steps_related if step not in st.session_state.selected_steps]
+    # Add new steps to the operations list.
+    st.session_state.selected_steps += new_selected_steps
+
+
+def remove_step(step: str):
+    """Remove a step from the operations list."""
+    st.session_state.selected_steps.remove(step)
+    if step in st.session_state.selected_steps_table:
+        st.session_state.selected_steps_table.remove(step)
